Move some useGetThankYouUrl dependencies to the function
